clean up formatted JSON diff
Before comparing JSON values, we wrap them in an object. We do so because gojsondiff/formatter only supports objects and arrays, while we also want to support values of basic types like ints and strings. That said, there should be no reason to expose the wrapping object to the user, so I added some code to remove it from the formatted diff.

I also added some test cases and an example.